Agregar la función de recuperación de contraseña para los usuarios
Implementa la recuperación de contraseñas por correo electrónico mediante Nodemailer, incluyendo DTOs para contraseña olvidada/restablecida, nuevos endpoints en users.controller y campos compatibles en la entidad de usuario. Además, actualiza la lógica de eliminación de cursos y módulos para eliminar en cascada las entidades relacionadas y mejora la gestión de archivos en el servicio leccion.

+++ src/cursos/cursos.service.ts
@@ -4,13 +4,16 @@ import { Repository } from 'typeorm';
 import { Curso } from './entities/curso.entity';
 import { CreateCursoDto } from './dto/create-curso.dto';
 import { UpdateCursoDto } from './dto/update-curso.dto';
+import { ModulosService } from 'src/modulos/modulos.service';
 
 @Injectable()
 export class CursosService {
   constructor(
     @InjectRepository(Curso)
     private readonly cursosRepository: Repository<Curso>,
-  ) {}
+
+    private readonly modulosService: ModulosService,
+  ) { }
 
   async create(createCursoDto: CreateCursoDto): Promise<Curso> {
     const curso = this.cursosRepository.create(createCursoDto);
@@ -26,6 +29,7 @@ export class CursosService {
       where: { id },
       relations: ['modulos'],
     });
+
     if (!curso) throw new NotFoundException('Curso not found');
     return curso;
   }
@@ -38,6 +42,15 @@ export class CursosService {
 
   async remove(id: number): Promise<void> {
     const curso = await this.findOne(id);
+
+    if (!curso) throw new NotFoundException('Curso not found');
+
+    const modulos = curso.modulos || [];
+
+    for (const modulo of modulos) {
+      await this.modulosService.remove(modulo.id);
+    }
+
     await this.cursosRepository.remove(curso);
   }
 }
